requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
PyYAML==5.3.1
requests==2.24.0
urllib3==1.25.9
